ajout de la fonction placerEvenementsEnColonne pour positionner les événements en colonne dans la vue hebdomadaire ajout du détail des événements dans la vue mensuelle
